Let the jmpress dependency to be loaded outside the angular wrapper
The only requirement is that the file is mapped as "jquery.jmpress".

This is necessary because one may want to use specific jmpress modules
with a custom build instead of the whole build or a pre-determined one.

Also, one may want to create a custom jmpress plugin and use it along
with angular wrapper.